restructured and cleaned of code
