--- core/src/Connections/sendPossition.java
@@ -1,12 +1,0 @@
-package Connections;
-
-/**
- * Created by Antonio M on 08/03/2018.
- */
-
-public class sendPossition implements Runnable {
-    @Override
-    public void run() {
-
-    }
-}
